Add explicit casting of arguments as double for pow in C++ examples. All examples now compile with the single precision version of the library.
svn path=/trunk/; revision=5813

+++ examples/c++/x01.cc
@@ -241,7 +241,7 @@ void x01::plot1( int do_test )
   for( i=0; i < 60; i++ )
     {
       x[i] = xoff + xscale * (i + 1) / 60.0;
-      y[i] = yoff + yscale * pow(x[i], 2.);
+      y[i] = yoff + yscale * pow((double)x[i], 2.);
     }
 
   xmin = x[0];
