+++ examples/c++/x22.cc
@@ -173,13 +173,13 @@ x22::potential() {
     d1 = rmax/4.;
 
     q1i = - q1*rmax/d1;
-    d1i = pow(rmax, 2.)/d1;
+    d1i = pow((double)rmax, 2.)/d1;
 
     q2 = -1.;
     d2 = rmax/4.;
 
     q2i = - q2*rmax/d2;
-    d2i = pow(rmax, 2.)/d2;
+    d2i = pow((double)rmax, 2.)/d2;
 
     for (i = 0; i < nr; i++) {
 	r = 0.5 + (double) i;
@@ -189,15 +189,15 @@ x22::potential() {
 	    y = r*sin(theta);
 	    cgrid2.xg[i][j] = x;
 	    cgrid2.yg[i][j] = y;
-	    div1 = sqrt(pow(x-d1, 2.) + pow(y-d1, 2.) + pow(eps, 2.));
-	    div1i = sqrt(pow(x-d1i, 2.) + pow(y-d1i, 2.) + pow(eps, 2.));
-	    div2 = sqrt(pow(x-d2, 2.) + pow(y+d2, 2.) + pow(eps, 2.));
-	    div2i = sqrt(pow(x-d2i, 2.) + pow(y+d2i, 2.) + pow(eps, 2.));
+	    div1 = sqrt(pow((double)(x-d1), 2.) + pow((double)(y-d1), 2.) + pow((double)eps, 2.));
+	    div1i = sqrt(pow((double)(x-d1i), 2.) + pow((double)(y-d1i), 2.) + pow((double)eps, 2.));
+	    div2 = sqrt(pow((double)(x-d2), 2.) + pow((double)(y+d2), 2.) + pow((double)eps, 2.));
+	    div2i = sqrt(pow((double)(x-d2i), 2.) + pow((double)(y+d2i), 2.) + pow((double)eps, 2.));
 	    z[i][j] = q1/div1 + q1i/div1i + q2/div2 + q2i/div2i;
-	    u[i][j] = -q1*(x-d1)/pow(div1,3.) - q1i*(x-d1i)/pow(div1i,3.0) 
-		- q2*(x-d2)/pow(div2,3.) - q2i*(x-d2i)/pow(div2i,3.);
-	    v[i][j] = -q1*(y-d1)/pow(div1,3.) - q1i*(y-d1i)/pow(div1i,3.0) 
-		- q2*(y+d2)/pow(div2,3.) - q2i*(y+d2i)/pow(div2i,3.);
+	    u[i][j] = -q1*(x-d1)/pow((double)div1,3.) - q1i*(x-d1i)/pow((double)div1i,3.0) 
+		- q2*(x-d2)/pow((double)div2,3.) - q2i*(x-d2i)/pow((double)div2i,3.);
+	    v[i][j] = -q1*(y-d1)/pow((double)div1,3.) - q1i*(y-d1i)/pow((double)div1i,3.0) 
+		- q2*(y+d2)/pow((double)div2,3.) - q2i*(y+d2i)/pow((double)div2i,3.);
 	}
     }
